+++ src/pf_imgui/elements/plots/types/Digital.cpp
@@ -8,6 +8,6 @@ namespace pf::ui::ig::plot_type {
 
 Digital::Digital(const std::string &elementName, const std::string &caption) : LabeledPlotData(elementName, caption) {}
 
-void Digital::renderImpl() { ImPlot::PlotDigital(getLabel().c_str(), xData.data(), yData.data(), xData.size()); }
+void Digital::renderImpl() { ImPlot::PlotDigital(getLabel().c_str(), xData.data(), yData.data(), static_cast<int>(xData.size())); }
 
 }// namespace pf::ui::ig::plot_type
